nba-news-frontend/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
alembic
requests
beautifulsoup4
python-dotenv
